+++ src/Richmat/WiLinke/controllerBuilder.ts
@@ -1,5 +1,4 @@
 import { IDeviceData } from '@ha/IDeviceData';
-import { logInfo } from '@utils/logger';
 import { BLEController } from 'BLE/BLEController';
 import { IBLEDevice } from 'ESPHome/types/IBLEDevice';
 import { supportedBeds } from './supportedBeds';
@@ -7,7 +6,7 @@ import { supportedBeds } from './supportedBeds';
 const buildCommand = (command: number) => [110, 1, 0, command, command + 111];
 
 export const controllerBuilder = async (deviceData: IDeviceData, bleDevice: IBLEDevice) => {
-  const { name, getCharacteristic } = bleDevice;
+  const { getCharacteristic } = bleDevice;
 
   for (const { serviceUuid, writeCharacteristicUuid } of supportedBeds) {
     const characteristic = await getCharacteristic(serviceUuid, writeCharacteristicUuid, false);
@@ -15,6 +14,5 @@ export const controllerBuilder = async (deviceData: IDeviceData, bleDevice: IBLE
     return new BLEController(deviceData, bleDevice, characteristic.handle, buildCommand);
   }
 
-  logInfo('[Richmat] Could not find any supported services/characteristics for device:', name);
   return undefined;
 };